requirements.txt
# Core dependencies for badger cropping pipeline
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.60.0

# SpeciesNet (CameraTrapAI) for badger detection
cameratrapai>=4.0.0

# Additional dependencies
matplotlib>=3.3.0 